crabot-memory/pyproject.toml
[project]
name = "crabot-memory"
version = "0.1.0"
description = "Crabot Memory Module"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "lancedb>=0.5.0",
    "pyarrow>=15.0.0",
    "openai>=1.12.0",
    "anthropic>=0.30.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0",
    "rank-bm25>=0.2.2",
    "aiosqlite>=0.19",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
memory = "src.main:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
